Переместит код из main в отдельные presenter
