+++ src/presenter/main-presenter.js
@@ -1,0 +1,6 @@
+import FilmsPresenter from '../presenter/films-presenter.js';
+
+const siteMain = document.querySelector('.main');
+const filmsPresenter = new FilmsPresenter({ filmsContainer: siteMain });
+
+filmsPresenter.init();
