+++ src/presenter/popup-presenter.js
@@ -1,0 +1,8 @@
+import FilmPopupView from '../view/film-popup-view.js';
+import { render } from '../render.js';
+
+const popupComponent = new FilmPopupView();
+const siteBody = document.body;
+
+
+render(popupComponent, siteBody);
